requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
#psycopg2-binary==2.9.9  # Для PostgreSQL
python-dotenv==1.0.0